pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clia-swe-ai"
version = "0.1.2"
description = "A CLI AI assistant for software engineering tasks."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.8"
license = { text = "MIT License" }
keywords = ["cli", "ai", "software-engineering", "gemini", "mcp"]
authors = [
  { name = "ayanda", email = "past952@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "google-genai",
    "mcp",
    "rich",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "psutil",
    "groq",
]

[project.urls]
"Homepage" = "https://github.com/Past-da-king/clia.git"
"Bug Tracker" = "https://github.com/Past-da-king/clia.git/issues"

[project.scripts]
clia = "gui.main:run_clia"

[tool.setuptools.packages.find]
where = ["."]
include = ["gui", "swe_tools", "core"]
namespaces = false
